requirements.txt
pydantic>=2.0
httpx>=0.27
pyyaml>=6.0
python-dotenv>=1.0
rich>=13.0
prompt_toolkit>=3.0
thefuzz>=0.22